backend/requirements.txt
python-dotenv
fastapi
uvicorn
neo4j
pandas
requests
openpyxl
uvloop          # WhisperServer